requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
colorama==0.4.6
distro==1.9.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.10
jiter==0.8.0
openai==1.56.1
pydantic==2.10.3
pydantic_core==2.27.1
python-dotenv==1.0.1
python-telegram-bot==20.3
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.12.2
